fix: bg-image + minimap bugs from Fable 5 hunt
Background image:
- HIGH: an image larger than the GPU max texture size (default 8192) was a
  validation error → wgpu's default handler panics the process (a wallpaper /
  panorama at startup or on hot-reload killed the terminal). load_background now
  downscales to device.limits().max_texture_dimension_2d.
- MEDIUM: opacity was forced to 1.0 unless the surface was PreMultiplied, so a
  background image was invisible on opaque (X11) surfaces even though it composites
  in-pass. Opacity now applies when translucent OR a background is set.
- LOW: w*h*4 buffer-size check computed in u32 (overflow) → now usize.
- LOW: hot-reload re-decoded the image on every config save; now cached by
  (path, dim) and only reloaded on a real change.

Minimap:
- HIGH: minimap_jump resolved the pane rect via layout() (split rects) while the
  strip is drawn from visible_rects() (zoom-aware), so clicks mapped to the wrong
  region when a pane was zoomed. Both use visible_rects now.
- MED: a pane narrower than the 46px strip had its hit-region and bars escape the
  pane; the minimap is now suppressed (draw and click) when r.w <= strip_w.
- The minimap click is handled after copy-mode exit, so a jump leaves copy-mode
  first.

Cover-crop math and jump math verified correct by the hunt. 163 tests.

+++ src/render.rs
@@ -1352,7 +1352,7 @@ impl Background {
             self.bind = None;
             return;
         };
-        if w == 0 || h == 0 || pixels.len() < (w * h * 4) as usize {
+        if w == 0 || h == 0 || pixels.len() < (w as usize) * (h as usize) * 4 {
             self.bind = None;
             return;
         }
